feat: support setting a default story type
Expose a `org-clubhouse-default-story-type` variable that can be set to
`feature`, `bug`, `chore` or `prompt`. If set to `prompt`, the
org-clubhouse `Create new story` flow will prompt for the story type
each time. If set to any of the other values, the `Create new story`
flow will use that value for all new stories until that variable is set
otherwise.

This also exposes a new interactive function called
`org-clubhouse-set-default-story-type` that prompts users to set the
default story type value.